+++ forge/src/main/java/com/telepathicgrunt/the_bumblezone/mixins/forge/item/ItemExtensionMixin.java
@@ -3,8 +3,10 @@ package com.telepathicgrunt.the_bumblezone.mixins.forge.item;
 import com.telepathicgrunt.the_bumblezone.platform.ItemExtension;
 import com.telepathicgrunt.the_bumblezone.utils.OptionalBoolean;
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.enchantment.Enchantment;
+import net.minecraft.world.level.Level;
 import net.minecraftforge.common.extensions.IForgeItem;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
@@ -24,6 +26,9 @@ public interface ItemExtensionMixin extends IForgeItem {
 
     @Shadow
     EquipmentSlot bz$getEquipmentSlot(ItemStack stack);
+
+    @Shadow
+    void bz$onArmorTick(ItemStack itemstack, Level world, Player player);
 
     @Override
     default int getMaxDamage(ItemStack stack) {
@@ -45,5 +50,10 @@ public interface ItemExtensionMixin extends IForgeItem {
     @Nullable
     default EquipmentSlot getEquipmentSlot(ItemStack stack) {
         return this.bz$getEquipmentSlot(stack);
+    }
+
+    @Override
+    default void onArmorTick(ItemStack itemstack, Level world, Player player) {
+        this.bz$onArmorTick(itemstack, world, player);
     }
 }