+++ src/main/java/com/telepathicgrunt/the_bumblezone/entities/mobs/BeehemothEntity.java
@@ -65,6 +65,7 @@ import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
 import net.minecraft.world.level.LevelReader;
+import net.minecraft.world.level.block.SweetBerryBushBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.phys.AABB;
 import net.minecraft.world.phys.HitResult;
@@ -215,6 +216,14 @@ public class BeehemothEntity extends TamableAnimal implements FlyingAnimal, Sadd
             return true;
         }
         return super.isInvulnerableTo(damageSource);
+    }
+
+    @Override
+    public void makeStuckInBlock(BlockState blockState, Vec3 speedMult) {
+        if (blockState.getBlock() instanceof SweetBerryBushBlock) {
+            return;
+        }
+        super.makeStuckInBlock(blockState, speedMult);
     }
 
     @Override